+++ Integral.Games/Statistics/DelegatedStatistic.cs
@@ -3,10 +3,10 @@ using Integral.Subscribers;
 
 namespace Integral.Statistics
 {
-    public sealed class DelegatedStatistic : ObservedStatistic, Subscriber<float>
+    public sealed class DelegatedStatistic : ObservedStatistic, Subscriber<int>
     {
-        public DelegatedStatistic(Publisher<float> publisher, float value = default) : base(value) => publisher.OnPublish += OnPublished;
+        public DelegatedStatistic(Publisher<int> publisher, int value = default) : base(value) => publisher.OnPublish += OnPublished;
 
-        public void OnPublished(float value) => Value = value;
+        public void OnPublished(int value) => Value = value;
     }
 }
